Restructure the scripts
